create TelegramApi trait
Allow to replace a default http client
