+++ src/lib.rs
@@ -5,3 +5,9 @@ pub mod objects;
 pub use api::*;
 pub use api_params::*;
 pub use objects::*;
+
+#[cfg(feature = "impl")]
+pub mod api_impl;
+
+#[cfg(feature = "impl")]
+pub use api_impl::*;
